Signup Actions: Remove `error` argument to `submitSignupStep`
While doing some unrelated work I realized with the help of @gwwar
that nothing is passing errors into the `submitSignupStep` function
and at the same time the places passing "no errors" were doing it
in three separate ways:
 - `undefined`
 - `null`
 - `[]`

In this patch I'm removing the error entirely and the corresponding
logic for handling it. If we need to pass an error someday we can
add it back in. It's probably not worth the prime second-position
in which it currently lives.